requirements.txt
cffi==1.9.1
crypto==1.4.1
cryptography==1.7.1
Django==1.10.5
django-cors-headers==2.0.0
djangorestframework==3.5.3
enum34==1.1.6
idna==2.2
ipaddress==1.0.18
Naked==0.1.31
oauthlib==2.0.1
py==1.4.32
pyasn1==0.1.9
pycparser==2.17
PyJWT==1.4.2
pyOpenSSL==16.2.0
pytest==3.0.5
pytest-rerunfailures==2.1.0
PyYAML==3.12
requests==2.12.4
requests-oauthlib==0.7.0
shellescape==3.4.1
six==1.10.0
TravisPy==0.3.5
